requirements.txt
djangorestframework
django-cors-headers
Pillow
djangorestframework-simplejwt
djoser
python-dotenv
drf-yasg